Rename OSX to macOS in the whole project
