+++ .github/workflows/ios.yml
@@ -1,4 +1,4 @@
-name: Package IOS
+name: Package iOS
 
 on:
   push:
@@ -59,5 +59,5 @@ jobs:
       - name: Export package
         uses: actions/upload-artifact@v4
         with:
-          name: package_IOS_${{ env.QBDI_ARCH }}
+          name: package_iOS_${{ env.QBDI_ARCH }}
           path: build/QBDI-*.tar.gz
