+++ CMakeLists.txt
@@ -221,7 +221,7 @@ if(QBDI_DISABLE_AVX)
 endif()
 
 # Force 32bit
-if(QBDI_PLATFORM_OSX AND QBDI_ARCH_X86)
+if(QBDI_PLATFORM_MACOS AND QBDI_ARCH_X86)
   set(CMAKE_C_FLAGS "-arch i386 ${CMAKE_C_FLAGS}")
   set(CMAKE_CXX_FLAGS "-arch i386 ${CMAKE_CXX_FLAGS}")
 elseif((QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID) AND QBDI_ARCH_X86)
@@ -260,7 +260,7 @@ elseif(QBDI_ARCH_X86 OR QBDI_ARCH_X86_64)
     enable_language(ASM)
   endif()
   set(ASM_EXT "s")
-  if(QBDI_PLATFORM_OSX AND QBDI_ARCH_X86)
+  if(QBDI_PLATFORM_MACOS AND QBDI_ARCH_X86)
     set(CMAKE_ASM-ATT_FLAGS "-arch i386")
   elseif((QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID) AND QBDI_ARCH_X86)
     set(CMAKE_ASM-ATT_FLAGS "--32")
