+++ README.rst
@@ -63,8 +63,8 @@ AArch64   Android, Linux, macOS, iOS (*)   :green:`Supported`   :green:`Supporte
 .. image:: https://github.com/QBDI/QBDI/actions/workflows/linux.yml/badge.svg?branch=master
     :target: https://github.com/QBDI/QBDI/actions/workflows/linux.yml?query=branch%3Amaster
 
-.. image:: https://github.com/QBDI/QBDI/actions/workflows/osx.yml/badge.svg?branch=master
-    :target: https://github.com/QBDI/QBDI/actions/workflows/osx.yml?query=branch%3Amaster
+.. image:: https://github.com/QBDI/QBDI/actions/workflows/macos.yml/badge.svg?branch=master
+    :target: https://github.com/QBDI/QBDI/actions/workflows/macos.yml?query=branch%3Amaster
 
 **dev**
 
@@ -78,8 +78,8 @@ AArch64   Android, Linux, macOS, iOS (*)   :green:`Supported`   :green:`Supporte
 .. image:: https://github.com/QBDI/QBDI/actions/workflows/linux.yml/badge.svg?branch=dev-next
     :target: https://github.com/QBDI/QBDI/actions/workflows/linux.yml?query=branch%3Adev-next
 
-.. image:: https://github.com/QBDI/QBDI/actions/workflows/osx.yml/badge.svg?branch=dev-next
-    :target: https://github.com/QBDI/QBDI/actions/workflows/osx.yml?query=branch%3Adev-next
+.. image:: https://github.com/QBDI/QBDI/actions/workflows/macos.yml/badge.svg?branch=dev-next
+    :target: https://github.com/QBDI/QBDI/actions/workflows/macos.yml?query=branch%3Adev-next
 
 .. intro-end
 
@@ -109,7 +109,7 @@ Devel packages can be downloaded in the artefacts of:
 
 - `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/android.yml?query=branch%3Adev-next>`__ for Android C/C++/frida API
 - `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/linux.yml?query=branch%3Adev-next>`__ for Linux C/C++/frida API (based on ubuntu)
-- `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/osx.yml?query=branch%3Adev-next>`__ for OSX C/C++/frida API
+- `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/macos.yml?query=branch%3Adev-next>`__ for macOS C/C++/frida API
 - `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/windows.yml?query=branch%3Adev-next>`__ for Windows packages C/C++ API
 - `GitHub Actions <https://github.com/QBDI/QBDI/actions/workflows/python.yml?query=branch%3Adev-next>`__ for PyQBDI
