+++ cmake/QBDIConfig.cmake
@@ -63,14 +63,14 @@ endif()
 # ========================
 # QBDI Platform
 # ========================
-set(QBDI_SUPPORTED_PLATFORMS "android" "linux" "windows" "osx" "ios")
+set(QBDI_SUPPORTED_PLATFORMS "android" "linux" "windows" "macos" "ios")
 
 set(QBDI_SUPPORTED_ARCH "ARM" "AARCH64" "X86" "X86_64")
 
 set(QBDI_PLATFORM_WINDOWS 0)
 set(QBDI_PLATFORM_LINUX 0)
 set(QBDI_PLATFORM_ANDROID 0)
-set(QBDI_PLATFORM_OSX 0)
+set(QBDI_PLATFORM_MACOS 0)
 set(QBDI_PLATFORM_IOS 0)
 
 set(QBDI_ARCH_X86 0)
@@ -94,7 +94,9 @@ elseif(QBDI_PLATFORM STREQUAL "linux")
 elseif(QBDI_PLATFORM STREQUAL "android")
   set(QBDI_PLATFORM_ANDROID 1)
 elseif(QBDI_PLATFORM STREQUAL "osx")
-  set(QBDI_PLATFORM_OSX 1)
+  set(QBDI_PLATFORM_MACOS 1)
+elseif(QBDI_PLATFORM STREQUAL "macos")
+  set(QBDI_PLATFORM_MACOS 1)
 elseif(QBDI_PLATFORM STREQUAL "ios")
   set(QBDI_PLATFORM_IOS 1)
 else()
@@ -142,7 +144,7 @@ if(QBDI_PLATFORM_WINDOWS AND NOT (QBDI_ARCH_X86 OR QBDI_ARCH_X86_64))
       "Windows platform is only supported for X86 and X86_64 architectures.")
 endif()
 
-if(QBDI_PLATFORM_OSX
+if(QBDI_PLATFORM_MACOS
    AND NOT
        (QBDI_ARCH_X86
         OR QBDI_ARCH_X86_64
@@ -158,7 +160,7 @@ if(QBDI_PLATFORM_IOS AND NOT (QBDI_ARCH_ARM OR QBDI_ARCH_AARCH64))
       "IOS platform is only supported for ARM and AARCH64 architecture.")
 endif()
 
-if(QBDI_PLATFORM_OSX)
+if(QBDI_PLATFORM_MACOS)
   if(QBDI_ARCH_X86_64 AND NOT ("${CMAKE_HOST_SYSTEM_PROCESSOR}" STREQUAL
                                "x86_64"))
     set(CMAKE_OSX_ARCHITECTURES
@@ -178,7 +180,7 @@ if(QBDI_PLATFORM_OSX)
   endif()
 endif()
 
-if(QBDI_PTRAUTH AND NOT ((QBDI_PLATFORM_OSX OR QBDI_PLATFORM_IOS)
+if(QBDI_PTRAUTH AND NOT ((QBDI_PLATFORM_MACOS OR QBDI_PLATFORM_IOS)
                          AND QBDI_ARCH_AARCH64))
   message(
     FATAL_ERROR
