+++ cmake/config/config-macOS-AARCH64.sh
@@ -7,6 +7,6 @@ GITDIR="$(realpath "${BASEDIR}/../..")"
 cmake "${GITDIR}"                        \
       -DCMAKE_BUILD_TYPE=Release         \
       -DCMAKE_CROSSCOMPILING=FALSE       \
-      -DQBDI_PLATFORM=osx                \
+      -DQBDI_PLATFORM=macos              \
       -DQBDI_ARCH=AARCH64                \
       -G Ninja
