+++ cmake/config/config-macOS-AARCH64e.sh
@@ -7,7 +7,7 @@ GITDIR="$(realpath "${BASEDIR}/../..")"
 cmake "${GITDIR}"                        \
       -DCMAKE_BUILD_TYPE=Release         \
       -DCMAKE_CROSSCOMPILING=FALSE       \
-      -DQBDI_PLATFORM=osx                \
+      -DQBDI_PLATFORM=macos              \
       -DQBDI_ARCH=AARCH64                \
       -DCMAKE_OSX_ARCHITECTURES="arm64e" \
       -DQBDI_PTRAUTH=ON                  \
