+++ cmake/config/config-macOS-X86.sh
@@ -7,6 +7,6 @@ GITDIR="$(realpath "${BASEDIR}/../..")"
 cmake "${GITDIR}"                        \
       -DCMAKE_BUILD_TYPE=Release         \
       -DCMAKE_CROSSCOMPILING=FALSE       \
-      -DQBDI_PLATFORM=osx                \
+      -DQBDI_PLATFORM=macos              \
       -DQBDI_ARCH=X86                    \
       -G Ninja
