+++ docs/source/get_started-PyQBDIPreload.rst
@@ -83,4 +83,4 @@ Apple silicon architecture
 
 In addition to the caveats discussed for QBDIPreload (see :ref:`macos-apple-silicon`), using PyQBDIPreload on ``arm64e`` binaries requires a Python interpreter built for the same ABI.
 
-Example commands for compiling OpenSSL and Python from sources for ``arm64e`` are provided in the workflow file ``.github/workflows/python_osx.yml``.
+Example commands for compiling OpenSSL and Python from sources for ``arm64e`` are provided in the workflow file ``.github/workflows/python_macos.yml``.
