+++ include/QBDI/PtrAuth.h
@@ -32,7 +32,7 @@ struct real_addr_t {
 
 #ifndef QBDI_ARCH_AARCH64
 #error "PTRAUTH is only implemented for AARCH64 architecture"
-#elif defined(QBDI_PLATFORM_OSX) || defined(QBDI_PLATFORM_IOS)
+#elif defined(QBDI_PLATFORM_MACOS) || defined(QBDI_PLATFORM_IOS)
 
 #include <ptrauth.h>
 namespace QBDI {
