+++ package/CMakeLists.txt
@@ -18,7 +18,7 @@ if(QBDI_PLATFORM_WINDOWS)
     set(CPACK_GENERATOR "NSIS")
   endif()
   set(CPACK_NSIS_PACKAGE_NAME "QBDI ${QBDI_VERSION_STRING} (${QBDI_ARCH})")
-elseif(QBDI_PLATFORM_OSX)
+elseif(QBDI_PLATFORM_MACOS)
   set(CPACK_GENERATOR "productbuild")
 else()
   set(CPACK_GENERATOR "TGZ")
