+++ src/ExecBlock/AARCH64/CMakeLists.txt
@@ -4,11 +4,11 @@ set(SOURCES "${CMAKE_CURRENT_LIST_DIR}/ExecBlock_AARCH64.cpp")
 if(QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID)
   list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/linux-android_AARCH64.s")
 
-elseif(QBDI_PLATFORM_OSX)
+elseif(QBDI_PLATFORM_MACOS)
   if(QBDI_PTRAUTH)
-    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/osx_AARCH64e.s")
+    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/macos_AARCH64e.s")
   else()
-    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/osx_AARCH64.s")
+    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/macos_AARCH64.s")
   endif()
 
 elseif(QBDI_PLATFORM_IOS)
