+++ src/ExecBlock/X86_64/CMakeLists.txt
@@ -1,11 +1,11 @@
 # Add QBDI target
 set(SOURCES "${CMAKE_CURRENT_LIST_DIR}/ExecBlock_X86_64.cpp")
 
-if(QBDI_PLATFORM_OSX)
+if(QBDI_PLATFORM_MACOS)
   if(QBDI_ARCH_X86_64)
-    set(ASM_STUB_EXECBLOCK "${CMAKE_CURRENT_LIST_DIR}/osx_X86_64.s")
+    set(ASM_STUB_EXECBLOCK "${CMAKE_CURRENT_LIST_DIR}/macos_X86_64.s")
   else() # QBDI_ARCH_X86
-    set(ASM_STUB_EXECBLOCK "${CMAKE_CURRENT_LIST_DIR}/osx_X86.s")
+    set(ASM_STUB_EXECBLOCK "${CMAKE_CURRENT_LIST_DIR}/macos_X86.s")
   endif()
   list(APPEND SOURCES "${ASM_STUB_EXECBLOCK}")
