+++ src/Utility/AARCH64/CMakeLists.txt
@@ -2,11 +2,11 @@ set(SOURCES "${CMAKE_CURRENT_LIST_DIR}/InstAnalysis_AARCH64.cpp")
 
 if(QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID)
   list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/linux-android-AARCH64.s")
-elseif(QBDI_PLATFORM_OSX)
+elseif(QBDI_PLATFORM_MACOS)
   if(QBDI_PTRAUTH)
-    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/osx-AARCH64e.s")
+    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/macos-AARCH64e.s")
   else()
-    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/osx-AARCH64.s")
+    list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/macos-AARCH64.s")
   endif()
   list(APPEND SOURCES "${CMAKE_CURRENT_LIST_DIR}/System_OSX.cpp"
        "${CMAKE_CURRENT_LIST_DIR}/System_darwin.cpp")
