+++ src/Utility/CMakeLists.txt
@@ -12,8 +12,9 @@ if(QBDI_PLATFORM_ANDROID OR QBDI_PLATFORM_LINUX)
   target_sources(
     QBDI_src INTERFACE "${CMAKE_CURRENT_LIST_DIR}/Memory_linux.cpp"
                        "${CMAKE_CURRENT_LIST_DIR}/System_generic.cpp")
-elseif(QBDI_PLATFORM_OSX OR QBDI_PLATFORM_IOS)
-  target_sources(QBDI_src INTERFACE "${CMAKE_CURRENT_LIST_DIR}/Memory_osx.cpp")
+elseif(QBDI_PLATFORM_MACOS OR QBDI_PLATFORM_IOS)
+  target_sources(QBDI_src
+                 INTERFACE "${CMAKE_CURRENT_LIST_DIR}/Memory_macos.cpp")
   if(NOT QBDI_ARCH_AARCH64)
     target_sources(QBDI_src
                    INTERFACE "${CMAKE_CURRENT_LIST_DIR}/System_generic.cpp")
