+++ src/Utility/X86_64/CMakeLists.txt
@@ -1,10 +1,10 @@
 set(SOURCES "${CMAKE_CURRENT_LIST_DIR}/InstAnalysis_X86_64.cpp")
 
-if(QBDI_PLATFORM_OSX)
+if(QBDI_PLATFORM_MACOS)
   if(QBDI_ARCH_X86_64)
-    set(ASM_STUB_UTILITY "${CMAKE_CURRENT_LIST_DIR}/osx-X86_64.s")
+    set(ASM_STUB_UTILITY "${CMAKE_CURRENT_LIST_DIR}/macos-X86_64.s")
   else() # QBDI_ARCH_X86
-    set(ASM_STUB_UTILITY "${CMAKE_CURRENT_LIST_DIR}/osx-X86.s")
+    set(ASM_STUB_UTILITY "${CMAKE_CURRENT_LIST_DIR}/macos-X86.s")
   endif()
   list(APPEND SOURCES "${ASM_STUB_UTILITY}")
