+++ test/API/AARCH64/CMakeLists.txt
@@ -6,7 +6,7 @@ target_sources(
           "${CMAKE_CURRENT_LIST_DIR}/VMTest_AARCH64_LocalMonitor.cpp"
           "${CMAKE_CURRENT_LIST_DIR}/VMTest_AARCH64_X28.cpp")
 
-if(NOT (QBDI_PLATFORM_OSX OR QBDI_PLATFORM_IOS))
+if(NOT (QBDI_PLATFORM_MACOS OR QBDI_PLATFORM_IOS))
   target_sources(QBDITest
                  PRIVATE "${CMAKE_CURRENT_LIST_DIR}/VMTest_AARCH64_LDRL.cpp")
 endif()
