+++ test/Patch/X86_64/Instr_Test_X86_64.cpp
@@ -58,7 +58,7 @@ TEST_CASE_METHOD(Instr_Test, "Instr_Test-UnalignedCodeBackward_IC") {
   INFO("Took " << count1 << " instructions\n");
 }
 
-#ifndef QBDI_PLATFORM_OSX
+#ifndef QBDI_PLATFORM_MACOS
 TEST_CASE_METHOD(Instr_Test, "Instr_Test-LoopCode_IC") {
   INFO("TEST_SEED=" << seed_random());
   uint64_t count1 = 0;
