+++ tools/validator/AARCH64/CMakeLists.txt
@@ -4,7 +4,7 @@ target_sources(validator
 if(QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID)
   target_sources(validator
                  PRIVATE "${CMAKE_CURRENT_LIST_DIR}/linux_AARCH64.cpp")
-elseif(QBDI_PLATFORM_OSX)
+elseif(QBDI_PLATFORM_MACOS)
   # no stub
 else()
   message(FATAL_ERROR "No stub for platform ${QBDI_PLATFORM}")
