+++ tools/validator/X86/CMakeLists.txt
@@ -3,7 +3,7 @@ target_sources(validator
 
 if(QBDI_PLATFORM_LINUX OR QBDI_PLATFORM_ANDROID)
   target_sources(validator PRIVATE "${CMAKE_CURRENT_LIST_DIR}/linux_X86.cpp")
-elseif(QBDI_PLATFORM_OSX)
+elseif(QBDI_PLATFORM_MACOS)
   # no stub
 else()
   message(FATAL_ERROR "No stub for platform ${QBDI_PLATFORM}")
